EventPersistenceSubscriber: PulseEventBus → Room
ViewModels: Today, Sleep, Activity, Coach with real DB data
Navigation: Settings route, Pairing route wired
TodayScreen: bound to TodayViewModel (real data)

Graph: 3246 nodes, 6574 edges (+37 from Phase 8)

+++ android/app/src/main/java/com/pulseloop/ring/RingDecodedEvent.kt
@@ -21,9 +21,9 @@ enum class SleepStage {
 
     companion object {
         fun fromByte(byte: UByte): SleepStage = when (byte) {
-            0x28u -> LIGHT
-            0x63u -> DEEP
-            0x00u -> AWAKE
+            0x28u.toUByte() -> LIGHT
+            0x63u.toUByte() -> DEEP
+            0x00u.toUByte() -> AWAKE
             else -> UNKNOWN
         }
     }
